+++ AWSS3/AWSS3TransferUtility.m
@@ -942,6 +942,7 @@ static AWSS3TransferUtility *_defaultS3TransferUtility = nil;
     AWSDDLogDebug(@"Value of timeoutIntervalForResource is %ld", (long)_transferUtilityConfiguration.timeoutIntervalForResource);
     getPreSignedURLRequest.minimumCredentialsExpirationInterval = _transferUtilityConfiguration.timeoutIntervalForResource;
     getPreSignedURLRequest.accelerateModeEnabled = self.transferUtilityConfiguration.isAccelerateModeEnabled;
+    getPreSignedURLRequest.preferredAccessStyle = self.transferUtilityConfiguration.preferredAccessStyle;
     
     [transferUtilityUploadTask.expression assignRequestHeaders:getPreSignedURLRequest];
     [transferUtilityUploadTask.expression assignRequestParameters:getPreSignedURLRequest];
@@ -1450,6 +1451,7 @@ internalDictionaryToAddSubTaskTo: (NSMutableDictionary *) internalDictionaryToAd
     request.expires = [NSDate dateWithTimeIntervalSinceNow:_transferUtilityConfiguration.timeoutIntervalForResource];
     request.minimumCredentialsExpirationInterval = _transferUtilityConfiguration.timeoutIntervalForResource;
     request.accelerateModeEnabled = self.transferUtilityConfiguration.isAccelerateModeEnabled;
+    request.preferredAccessStyle = self.transferUtilityConfiguration.preferredAccessStyle;
     [self filterAndAssignHeaders:transferUtilityMultiPartUploadTask.expression.requestHeaders getPresignedURLRequest:request
                       URLRequest:nil];
     
@@ -1662,6 +1664,7 @@ internalDictionaryToAddSubTaskTo: (NSMutableDictionary *) internalDictionaryToAd
     getPreSignedURLRequest.expires = [NSDate dateWithTimeIntervalSinceNow:_transferUtilityConfiguration.timeoutIntervalForResource];
     getPreSignedURLRequest.minimumCredentialsExpirationInterval = _transferUtilityConfiguration.timeoutIntervalForResource;
     getPreSignedURLRequest.accelerateModeEnabled = self.transferUtilityConfiguration.isAccelerateModeEnabled;
+    getPreSignedURLRequest.preferredAccessStyle = self.transferUtilityConfiguration.preferredAccessStyle;
     
     [transferUtilityDownloadTask.expression assignRequestHeaders:getPreSignedURLRequest];
     [transferUtilityDownloadTask.expression assignRequestParameters:getPreSignedURLRequest];
@@ -2593,6 +2596,7 @@ totalBytesExpectedToWrite:(int64_t)totalBytesExpectedToWrite {
         _retryLimit = 0;
         _multiPartConcurrencyLimit = @(AWSS3TransferUtilityMultiPartDefaultConcurrencyLimit);
         _timeoutIntervalForResource = AWSS3TransferUtilityTimeoutIntervalForResource;
+        _preferredAccessStyle = AWSS3BucketAccessStyleVirtualHosted;
     }
     return self;
 }
@@ -2604,6 +2608,7 @@ totalBytesExpectedToWrite:(int64_t)totalBytesExpectedToWrite {
     configuration.retryLimit = self.retryLimit;
     configuration.multiPartConcurrencyLimit = self.multiPartConcurrencyLimit;
     configuration.timeoutIntervalForResource = self.timeoutIntervalForResource;
+    configuration.preferredAccessStyle = self.preferredAccessStyle;
     return configuration;
 }